Modify M236 and setOutputPressure to use floating point number instead of integer for desired pressure.

+++ Marlin/Marlin_main.cpp
@@ -4149,7 +4149,7 @@ inline void gcode_M226() {
    */
   inline void gcode_M236() {
     if(code_seen('S')) {
-      uint8_t psi = code_value();
+      float psi = code_value();
       // Check that desired pressure is within range allowed
       if((psi <= OUTPUT_PSI_MAX) && (psi >= OUTPUT_PSI_MIN)) {
         setOutputPressure(psi);
